backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-JWT-Extended==4.5.2
Flask-Cors==4.0.1
python-dotenv==1.0.0
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
